Phase out i686: https://www.archlinux.org/news/the-end-of-i686-support/

+++ PKGBUILD
@@ -6,7 +6,7 @@ pkgname=nvidia-beta
 pkgver=387.22
 pkgrel=1
 pkgdesc="NVIDIA driver for Arch's official 'linux' package (beta version)"
-arch=('i686' 'x86_64')
+arch=('x86_64')
 url="http://www.nvidia.com/"
 license=('custom:NVIDIA')
 depends=('linux' "nvidia-utils-beta>=$pkgver" 'libgl')
@@ -15,20 +15,13 @@ provides=("nvidia=$pkgver")
 conflicts=('nvidia-96xx' 'nvidia-173xx' 'nvidia')
 options=('!strip')
 install=$pkgname.install
+_pkg="NVIDIA-Linux-x86_64-$pkgver-no-compat32"
+source=("http://us.download.nvidia.com/XFree86/Linux-x86_64/$pkgver/$_pkg.run")
+md5sums=('b21f9bafb20409b337505c9b1d362c34')
 
-# Installer name
-case "$CARCH" in
-  i686)   _pkg="NVIDIA-Linux-x86-$pkgver" ;;
-  x86_64) _pkg="NVIDIA-Linux-x86_64-$pkgver-no-compat32" ;;
-esac
-
-# Source
-#source=('linux-4.11.patch')
-source_i686=("http://us.download.nvidia.com/XFree86/Linux-x86/$pkgver/NVIDIA-Linux-x86-$pkgver.run")
-source_x86_64=("http://us.download.nvidia.com/XFree86/Linux-x86_64/$pkgver/NVIDIA-Linux-x86_64-$pkgver-no-compat32.run")
-#md5sums=('cc8941b6898d9daa0fb67371f57a56b6')
-md5sums_i686=('abdbb3c813e52148f7435308121e6a37')
-md5sums_x86_64=('b21f9bafb20409b337505c9b1d362c34')
+# Patch
+#source+=('linux-4.11.patch')
+#md5sums+=('cc8941b6898d9daa0fb67371f57a56b6')
 
 # Auto-detect patches (e.g. linux-4.1.patch)
 for _patch in $(find "$startdir" -maxdepth 1 -name '*.patch' -printf "%f\n"); do
@@ -41,9 +34,7 @@ done
 
 prepare() {
   # Remove previous builds
-  if [[ -d $_pkg ]]; then
-    rm -rf $_pkg
-  fi
+  [[ -d $_pkg ]] && rm -rf $_pkg
 
   # Extract
   msg2 "Self-Extracting $_pkg.run..."
@@ -55,7 +46,7 @@ prepare() {
     # Version variables
     _kernel=$(cat /usr/lib/modules/extramodules-*-ARCH/version)
     _major_patch=$(echo $_patch | grep -Po "\d+\.\d+")
-      
+
     # Check version
     if (( $(vercmp $_kernel $_major_patch) >= 0 )); then
       msg2 "Applying $_patch..."
@@ -86,10 +77,8 @@ package() {
          "$pkgdir"/usr/lib/modules/$_extramodules/nvidia.ko
 
   # Install UVM module: http://devblogs.nvidia.com/parallelforall/unified-memory-in-cuda-6/
-  if [[ $CARCH = x86_64 ]]; then
-    install -Dm644 $_pkg/kernel/nvidia-uvm.ko \
-           "$pkgdir"/usr/lib/modules/$_extramodules/nvidia-uvm.ko
-  fi
+  install -Dm644 $_pkg/kernel/nvidia-uvm.ko \
+         "$pkgdir"/usr/lib/modules/$_extramodules/nvidia-uvm.ko
 
   # Install Modeset module:
   #
